refactor pipelinerunner into stepsrunner and moduleloader
